+++ test/internals_test.rb
@@ -3,6 +3,7 @@ require_relative "helper"
 class TestInternals < Test::Unit::TestCase
 
   include Helper::Client
+  include Helper::Skipable
 
   def test_logger
     r.ping
@@ -156,6 +157,25 @@ class TestInternals < Test::Unit::TestCase
       Redis.new(opts).ping
     end
     assert (Time.now - start_time) <= opts[:timeout]
+  end
+
+  driver(:ruby) do
+    def test_write_timeout
+      return skip("Relies on buffer sizes, might be unreliable")
+
+      server = TCPServer.new("127.0.0.1", 0)
+      port   = server.addr[1]
+
+      # Hacky, but we need the buffer size
+      val = TCPSocket.new("127.0.0.1", port).getsockopt(Socket::SOL_SOCKET, Socket::SO_SNDBUF).unpack("i")[0]
+
+      assert_raise(Redis::TimeoutError) do
+        Timeout.timeout(1) do
+          redis = Redis.new(:port => port, :timeout => 5, :write_timeout => 0.1)
+          redis.set("foo", "1" * val*2)
+        end
+      end
+    end
   end
 
   def close_on_ping(seq, options = {})